Update 3
Most everything is working except the pitch gyro and I have no idea why.
The pages work but are no more than templates at this point and engine
control won't be coded till there are actually engines to test with and
that damn gyro is done.
Far in the future when ships become customizable, I will have to recode
a lot of this to make it flex to allow multiple sets of gyros and 2-set
vs. 3-set gyro systems, ect.